refactor crates/runtime/src/model from file to directory; PR comments
