+++ crates/runtime/src/http/routes.rs
@@ -15,8 +15,8 @@ limitations under the License.
 */
 
 use crate::embeddings::vector_search;
+use crate::model::EmbeddingModelStore;
 use crate::model::LLMModelStore;
-use crate::EmbeddingModelStore;
 use crate::{config, datafusion::DataFusion};
 use app::App;
 use axum::routing::patch;
